COMMERCE-5227 Move CommerceDiscountValidatorServiceWrapperPriorityComparator to comparator package
